% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastprimers2genome.R
\name{blastprimers2genome}
\alias{blastprimers2genome}
\title{blastprimers2genome: BLAST set of primer sequences against target genome.}
\usage{
blastprimers2genome(indir, genomepath, outdir, blastpath)
}
\arguments{
\item{indir}{Specify the path of the directory in which your primer sequences are stored.}

\item{genomepath}{Specifiy the path of the directory in which your blast database is stored.}

\item{outdir}{Specify the path of the directory in which the output file should be written.}

\item{blastpath}{String. Specify the path of the directory in which the blastn executable is.}
}
\description{
blastprimers2genome: BLAST set of primer sequences against target genome.
}

