% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastseqs2genome.R
\name{blastseqs2genome}
\alias{blastseqs2genome}
\title{blastseqs2genome: BLAST a query sequence against target genome.}
\usage{
blastseqs2genome(indir, genomepath, outdir, blastpath)
}
\arguments{
\item{indir}{String. Specify the path of the directory in which your query sequences are stored.}

\item{genomepath}{String. Specifiy the path of the directory in which your blast database is stored.}

\item{outdir}{String. Specify the path of the directory in which the output file should be written.}

\item{blastpath}{String. Specify the path of the directory in which the blastn executable is.}
}
\description{
blastseqs2genome: BLAST a query sequence against target genome.
}

