% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c.lupus_mc1r_seq}
\alias{c.lupus_mc1r_seq}
\title{Canis lupus familiaris melanocortin 1 receptor (alpha melanocyte stimulating hormone receptor) (MC1R), mRNA}
\format{A  list of 1 vector of chars, class: SeqFastadna}
\source{
\url{http://www.ncbi.nlm.nih.gov/nuccore/NM_001014282.2}
}
\usage{
c.lupus_mc1r_seq
}
\description{
Canis lupus familiaris melanocortin 1 receptor (alpha melanocyte stimulating hormone receptor) (MC1R), mRNA
}
\keyword{datasets}

