% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_primer3_input.R
\name{create_primer3_input_C}
\alias{create_primer3_input_C}
\title{create_primer3_input_C: Create the input file for running primer3 using the primer3 v.0.4.0 web parameters}
\usage{
create_primer3_input_C(seqID, target, primer3path, outdir)
}
\arguments{
\item{seqID}{Specify a name for your sequence.}

\item{target}{Specify your target sequence object.}

\item{primer3path}{Specify the path to the program primer3 for example: /home/emily/programs/primer3-2.3.7}

\item{outdir}{Specify the path of the directory in which the output file should be written.}
}
\description{
create_primer3_input_C: Create the input file for running primer3 using the primer3 v.0.4.0 web parameters
}

