% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_query_seqs.R
\name{download_query_seqs}
\alias{download_query_seqs}
\title{download_query_seqs: Download query sequences from GenBank}
\usage{
download_query_seqs(accessions, output)
}
\arguments{
\item{accessions}{Specify the GenBank accession number or numbers you would like to download.}

\item{output}{Specify the path and name of the output file (fasta format)}
}
\description{
download_query_seqs: Download query sequences from GenBank
}

