% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_target_location.R
\name{extract_target_location}
\alias{extract_target_location}
\title{extract_target_location: Extract the target location of the genome as a fasta sequence.}
\usage{
extract_target_location(chr, start, end, outdir, bedtoolspath, genomepath)
}
\arguments{
\item{chr}{Specify the scaffold, contig or chromosome in which gene is located.}

\item{start}{Specify the start coordinate of the sequence to be extracted.}

\item{end}{Specify the end coordinate of the sequence to be extracted.}

\item{outdir}{Specify the path of the directory in which the output file should be written.}

\item{bedtoolspath}{Specifiy the path of the directory in which the bedtools executable is stored.}

\item{genomepath}{Specifiy the path of the directory in which your blast database is stored.}
}
\description{
extract_target_location: Extract the target location of the genome as a fasta sequence.
}

