% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_str.R
\name{insert_str}
\alias{insert_str}
\title{insert_str: Insert characters into string at given index
https://goo.gl/tnCyQ9 & http://goo.gl/fPNaJO}
\usage{
insert_str(target, insert, index)
}
\arguments{
\item{target}{string to be edited}

\item{insert}{character to insert}

\item{index}{location to insert}
}
\description{
insert_str: Insert characters into string at given index
https://goo.gl/tnCyQ9 & http://goo.gl/fPNaJO
}

