% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_primer3.R
\name{run_primer3}
\alias{run_primer3}
\title{run_primer3: Run primer3_core executable}
\usage{
run_primer3(input, primer3path, indir, outdir)
}
\arguments{
\item{input}{Specify the name of your primer3 input file.}

\item{primer3path}{Specify the path of where the primer3_core executable is stored.}

\item{indir}{Specify the path of the directory in which the input file is stored.}

\item{outdir}{Specify the path of the directory in which the output file should be written.}
}
\description{
run_primer3: Run primer3_core executable
}

