% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_str.R
\name{split_str_by_index}
\alias{split_str_by_index}
\title{split_str_by_index}
\usage{
split_str_by_index(target, index)
}
\arguments{
\item{target}{string to edit}

\item{index}{location to edit}
}
\description{
split_str_by_index
}

