% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u.americanus_mc1r_primers}
\alias{u.americanus_mc1r_primers}
\title{Primer sequences used by Ritland et al. (2001) to sequence the MC1R in the Kermode Bear}
\format{A  list of 1 vector of chars, class: SeqFastadna}
\source{
\url{http://www.sciencedirect.com/science/article/pii/S0960982201004481}
}
\usage{
u.americanus_mc1r_primers
}
\description{
Primer sequences used by Ritland et al. (2001) to sequence the MC1R in the Kermode Bear
}
\keyword{datasets}

